fix(defaults): fix image defaults logic
- Moved apply_image_defaults function to config.rs
- Added null_to_none helper in db.rs to handle SQL null values
- Updated sandbox.rs to use the moved function and handle image defaults earlier
- Added more debug logging for configuration application
- Fixed shell handling in script generation
